fixed argument type.

+++ lib/output.c
@@ -255,7 +255,7 @@ grn_output_uint64(grn_ctx *ctx, grn_obj *outbuf, grn_content_type output_type, i
 }
 
 void
-grn_output_float(grn_ctx *ctx, grn_obj *outbuf, grn_content_type output_type, int64_t value)
+grn_output_float(grn_ctx *ctx, grn_obj *outbuf, grn_content_type output_type, double value)
 {
   put_delimiter(ctx, outbuf, output_type);
   switch (output_type) {
